added slugs in sub categories

+++ controller/blogs.controller.js
@@ -189,7 +189,7 @@ async function get(req, res) {
         b.updated_at,
         bc.category 
       FROM blogs b 
-      LEFT JOIN blogs_category bc ON bc.id::integer = b.category::integer
+      LEFT JOIN blogs_category bc ON bc.id = b.category::integer
       ORDER BY b.created_at DESC;`);
     res.json(rows);
   } catch (error) {
